backend done

+++ database/migrations/2025_10_14_161413_create_account_employee_table.php
@@ -8,20 +8,22 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('account_employee', function (Blueprint $table) {
-            $table->unsignedInteger('account_id');
-            $table->unsignedInteger('employee_id');
+        if (!Schema::hasTable('account_employee')) {
+            Schema::create('account_employee', function (Blueprint $table) {
+                $table->unsignedInteger('account_id');
+                $table->unsignedInteger('employee_id');
 
-            $table->foreign('account_id')
-                ->references('id')->on('account')
-                ->onDelete('cascade');
+                $table->foreign('account_id')
+                    ->references('id')->on('account')
+                    ->onDelete('cascade');
 
-            $table->foreign('employee_id')
-                ->references('id')->on('employee')
-                ->onDelete('cascade');
+                $table->foreign('employee_id')
+                    ->references('id')->on('employee')
+                    ->onDelete('cascade');
 
-            $table->primary(['account_id', 'employee_id']);
-        });
+                $table->primary(['account_id', 'employee_id']);
+            });
+        }
     }
 
     public function down(): void
